Remove min and max temperature for CurrentTemperature
The Min and Max values should not be set for CurrentTemperature, but only for TargetTemperature. If you set it, and the real temperature is under or above thies values, Home will not show the real temperature in the room.
You can also reduce the step to 0.1 for this temperature to be more accurate.